Extract Service generic class
